+++ Tests/UIKitTests/UIButtonExtensionsTests.swift
@@ -152,10 +152,33 @@ final class UIButtonExtensionsTests: XCTestCase {
     }
 
     func testCenterTextAndImage() {
-        let button = UIButton()
-        button.centerTextAndImage(spacing: 20)
-        XCTAssertEqual(button.imageEdgeInsets, UIEdgeInsets(top: 0, left: -10, bottom: 0, right: 10))
-        XCTAssertEqual(button.titleEdgeInsets, UIEdgeInsets(top: 0, left: 10, bottom: 0, right: -10))
+        let button = UIButton(frame: CGRect(x: 10, y: 10, width: 100, height: 100))
+        let image = UIImage(color: .green, size: CGSize(width: 10, height: 10))
+        button.setTitleForAllStates("Title")
+        button.setImageForAllStates(image)
+
+        XCTAssertNotNil(button.imageView)
+        XCTAssertNotNil(button.titleLabel)
+
+        let spacing: CGFloat = 20
+
+        // Image on the left of text
+        button.centerTextAndImage(spacing: spacing)
+        var imageFrame = button.imageView!.frame
+        var titleFrame = button.titleLabel!.frame
+
+        XCTAssert(titleFrame.midX > imageFrame.midX)
+        XCTAssertEqual(titleFrame.midY, imageFrame.midY, accuracy: 1.0)
+        XCTAssertEqual(titleFrame.minX - spacing, imageFrame.maxX, accuracy: 1.0)
+
+        // Image above text
+        button.centerTextAndImage(imageAboveText: true, spacing: spacing)
+        imageFrame = button.imageView!.frame
+        titleFrame = button.titleLabel!.frame
+
+        XCTAssert(titleFrame.midY > imageFrame.midY)
+        XCTAssertEqual(titleFrame.midX, imageFrame.midX, accuracy: 1.0)
+        XCTAssertEqual(titleFrame.minY - spacing, imageFrame.maxY, accuracy: 1.0)
     }
 
 }